Added LockWindowManageable protocol
